\name{trt}
\alias{trt}
\title{Mark Treatment Term}
\description{This is a dummy function, used to mark the treatment term in various functions within the \code{uplift} package.
}

\usage{
trt(x)
}

\arguments{
  \item{x}{A numeric variable coded as 1 (treatment) and 0 (control).
}
}

\value{
x, unchanged
}

\author{
Leo Guelman <leo.guelman@gmail.com>
}
