\name{uplift-package}
\alias{uplift-package}
\docType{package}
\title{Uplift Modeling
}
\description{An integrated package for building and testing uplift models.
}
\details{
\tabular{ll}{
Package: \tab uplift\cr
Type: \tab Package\cr
Version: \tab 0.3.4\cr
Date: \tab 2014-01-03\cr
Depends: \tab R (>= 3.0.0), RItools, MASS, coin, tables, penalized\cr
License: \tab  GPL-2 | GPL-3\cr
}
}
\author{
Leo Guelman <leo.guelman@gmail.com>
}

\keyword{package}
